202408-chap3-langchain-translate/requirements.txt
langchain==0.2.3
langchain-openai==0.1.8
streamlit==1.35.0
